+++ docs/whats-new.md
@@ -7,6 +7,7 @@
 
 ### Fix
 - improved performance of `dfmt.merge_meteofiles()` by adding xarray arguments and ensuring alignment in [#1148](https://github.com/Deltares/dfm_tools/pull/1148)
+- allow for number coordinate variable in `dfmt.merge_meteofiles()` in [#1157](https://github.com/Deltares/dfm_tools/pull/1157)
 - proper time slicing and out of bounds error of `dfmt.merge_meteofiles()` in [#1149](https://github.com/Deltares/dfm_tools/pull/1149)